function/requirements.txt
functions-framework==3.0.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.15.0
google-cloud-pubsub==2.19.0
